+++ src/main/java/craftedMods/recipes/api/ItemHidingHandler.java
@@ -1,0 +1,11 @@
+package craftedMods.recipes.api;
+
+import java.util.Collection;
+
+import net.minecraft.item.ItemStack;
+
+public interface ItemHidingHandler {
+	
+	public Collection<ItemStack> getHiddenStacks();
+
+}
