+++ src/main/java/craftedMods/recipes/provider/NEIIntegrationManager.java
@@ -1,11 +1,13 @@
 package craftedMods.recipes.provider;
 
-import java.util.List;
+import java.lang.annotation.Annotation;
+import java.util.*;
 
 import org.apache.logging.log4j.Logger;
 
 import codechicken.nei.api.API;
 import codechicken.nei.recipe.*;
+import craftedMods.recipes.NEIExtensions;
 import craftedMods.recipes.api.*;
 import craftedMods.utils.ClassDiscoverer;
 import lotr.common.LOTRMod;
@@ -21,6 +23,8 @@ public class NEIIntegrationManager {
 
 	private RecipeHandlerManager recipeHandlerManager;
 
+	private Collection<ItemHidingHandler> itemHidingHandlers = new ArrayList<>();
+
 	public NEIIntegrationManager(NEIExtensionsConfiguration config, Logger logger) {
 		this.config = config;
 		this.logger = logger;
@@ -30,6 +34,7 @@ public class NEIIntegrationManager {
 	public void preInit() {
 		this.discoverer.registerClassToDiscover(RegisteredHandler.class, RecipeHandler.class);
 		this.discoverer.registerClassToDiscover(RegisteredHandler.class, RecipeHandlerFactory.class);
+		this.discoverer.registerClassToDiscover(RegisteredHandler.class, ItemHidingHandler.class);
 		this.discoverer.discoverClassesAsync();
 	}
 
@@ -37,15 +42,37 @@ public class NEIIntegrationManager {
 		try {
 			long start = System.currentTimeMillis();
 
-			this.recipeHandlerManager = new RecipeHandlerManager(this.config.getConfigFile(),
-					this.discoverer.getDiscoveredClasses(this.config.getClassDiscovererThreadTimeout()));
+			Map<Class<? extends Annotation>, Map<Class<?>, Set<Class<?>>>> discoveredClasses = this.discoverer
+					.getDiscoveredClasses(this.config.getClassDiscovererThreadTimeout());
+
+			this.recipeHandlerManager = new RecipeHandlerManager(this.config.getConfigFile(), discoveredClasses);
 
 			this.recipeHandlerManager.init(useCachedRecipes);
+
+			NEIExtensions.mod.getLogger().info("Enable item hiding handlers: " + this.config.isHideTechnicalBlocks());
+			if (this.config.isHideTechnicalBlocks()) this.discoverItemHidingHandlers(discoveredClasses);
 
 			this.logger.info("Initialized NEI config for LOTR Mod within " + (System.currentTimeMillis() - start) + " ms");
 		} catch (Exception e) {
 			this.logger.error("Couldn't initialize NEI config for LOTR Mod: ", e);
 		}
+	}
+
+	@SuppressWarnings("unchecked")
+	private void discoverItemHidingHandlers(Map<Class<? extends Annotation>, Map<Class<?>, Set<Class<?>>>> discoveredClasses) {
+		Set<Class<?>> itemHidingHandlers = discoveredClasses.get(RegisteredHandler.class).get(ItemHidingHandler.class);
+		NEIExtensions.mod.getLogger().info("Found " + itemHidingHandlers.size() + " item hiding handlers in classpath");
+		itemHidingHandlers.forEach(clazz -> {
+			try {
+				Class<? extends ItemHidingHandler> handler = (Class<? extends ItemHidingHandler>) clazz;
+				if (handler.getAnnotation(RegisteredHandler.class).isEnabled()) {
+					this.itemHidingHandlers.add(handler.newInstance());
+					NEIExtensions.mod.getLogger().debug("Successfully registered item hiding handler \"" + handler.getName() + "\"");
+				} else NEIExtensions.mod.getLogger().info("The item hiding handler \"" + handler.getName() + "\" was disabled by the author.");
+			} catch (Exception e) {
+				NEIExtensions.mod.getLogger().error("Couldn't create an instance of class \"" + clazz.getName() + "\"", e);
+			}
+		});
 	}
 
 	public void load() {
@@ -63,7 +90,7 @@ public class NEIIntegrationManager {
 			this.recipeHandlerManager.getRecipeHandlers().forEach((unlocalizedName, handler) -> this.loadHandler(new PluginRecipeHandler<>(handler)));
 
 			// Item hiding
-			if (this.config.isHideTechnicalBlocks()) this.hideItems();
+			if (this.config.isHideTechnicalBlocks()) this.registerHiddenItems();
 
 			// Override names
 			this.addOverrideNames();
@@ -101,130 +128,15 @@ public class NEIIntegrationManager {
 		}
 	}
 
-	private void hideItems() {
-		// Beds
-		this.hideItemAll(LOTRMod.dwarvenBed);
-		this.hideItemAll(LOTRMod.elvenBed);
-		this.hideItemAll(LOTRMod.highElvenBed);
-		this.hideItemAll(LOTRMod.lionBed);
-		this.hideItemAll(LOTRMod.orcBed);
-		this.hideItemAll(LOTRMod.strawBed);
-		this.hideItemAll(LOTRMod.woodElvenBed);
-		this.hideItemAll(LOTRMod.furBed);
-
-		// Cakes/Pies
-		this.hideItemAll(LOTRMod.berryPie);
-		this.hideItemAll(LOTRMod.cherryPie);
-		this.hideItemAll(LOTRMod.dalishPastry);
-		this.hideItemAll(LOTRMod.appleCrumble);
-		this.hideItemAll(LOTRMod.bananaCake);
-		this.hideItemAll(LOTRMod.lemonCake);
-
-		// Fruits
-		this.hideItemAll(LOTRMod.dateBlock);
-		this.hideItemAll(LOTRMod.bananaBlock);
-		this.hideItemAll(LOTRMod.grapevineRed);
-		this.hideItemAll(LOTRMod.grapevineWhite);
-
-		// Torches
-		this.hideItemAll(LOTRMod.orcTorch);
-		this.hideItemAll(LOTRMod.tauredainDoubleTorch);
-
-		// Crops
-		this.hideItemAll(LOTRMod.lettuceCrop);
-		this.hideItemAll(LOTRMod.pipeweedCrop);
-		this.hideItemAll(LOTRMod.flaxCrop);
-		this.hideItemAll(LOTRMod.leekCrop);
-		this.hideItemAll(LOTRMod.turnipCrop);
-		this.hideItemAll(LOTRMod.yamCrop);
-
-		// Spawner Chests
-		this.hideItemAll(LOTRMod.spawnerChest);
-		this.hideItemAll(LOTRMod.spawnerChestStone);
-
-		// Vessels
-		this.hideItemAll(LOTRMod.mugBlock);
-		this.hideItemAll(LOTRMod.ceramicMugBlock);
-		this.hideItemAll(LOTRMod.gobletGoldBlock);
-		this.hideItemAll(LOTRMod.gobletSilverBlock);
-		this.hideItemAll(LOTRMod.gobletCopperBlock);
-		this.hideItemAll(LOTRMod.gobletWoodBlock);
-		this.hideItemAll(LOTRMod.skullCupBlock);
-		this.hideItemAll(LOTRMod.wineGlassBlock);
-		this.hideItemAll(LOTRMod.glassBottleBlock);
-		this.hideItemAll(LOTRMod.aleHornBlock);
-		this.hideItemAll(LOTRMod.aleHornGoldBlock);
-
-		// Others
-		this.hideItemAll(LOTRMod.flowerPot);
-		this.hideItemAll(LOTRMod.plateBlock);
-		this.hideItemAll(LOTRMod.armorStand);
-		this.hideItemAll(LOTRMod.marshLights);
-		this.hideItemAll(LOTRMod.utumnoReturnLight);
-		this.hideItemAll(LOTRMod.signCarved);
-		this.hideItemAll(LOTRMod.signCarvedIthildin);
-		this.hideItemAll(LOTRMod.bookshelfStorage);
-
-		// Slabs
-		this.hideItemMeta(LOTRMod.slabSingle);
-		this.hideItemMeta(LOTRMod.slabSingle2);
-		this.hideItemMeta(LOTRMod.slabSingle3);
-		this.hideItemMeta(LOTRMod.slabSingle4);
-		this.hideItemMeta(LOTRMod.slabSingle5);
-		this.hideItemMeta(LOTRMod.slabSingle6);
-		this.hideItemMeta(LOTRMod.slabSingle7);
-		this.hideItemMeta(LOTRMod.slabSingle8);
-		this.hideItemMeta(LOTRMod.slabSingle9);
-		this.hideItemMeta(LOTRMod.slabSingle10);
-		this.hideItemMeta(LOTRMod.slabSingle11);
-		this.hideItemMeta(LOTRMod.slabSingle12);
-		this.hideItemMeta(LOTRMod.slabSingleV);
-		this.hideItemMeta(LOTRMod.slabSingleThatch);
-		this.hideItemMeta(LOTRMod.slabSingleDirt);
-		this.hideItemMeta(LOTRMod.slabSingleSand);
-		this.hideItemMeta(LOTRMod.slabSingleGravel);
-		this.hideItemMeta(LOTRMod.rottenSlabSingle);
-		this.hideItemMeta(LOTRMod.scorchedSlabSingle);
-		this.hideItemMeta(LOTRMod.slabUtumnoSingle);
-		this.hideItemMeta(LOTRMod.slabUtumnoSingle2);
-		this.hideItemMeta(LOTRMod.slabClayTileSingle);
-		this.hideItemMeta(LOTRMod.slabClayTileDyedSingle);
-		this.hideItemMeta(LOTRMod.slabClayTileDyedSingle2);
-		this.hideItemMeta(LOTRMod.woodSlabSingle);
-		this.hideItemMeta(LOTRMod.woodSlabSingle2);
-		this.hideItemMeta(LOTRMod.woodSlabSingle3);
-		this.hideItemMeta(LOTRMod.woodSlabSingle4);
-		this.hideItemMeta(LOTRMod.woodSlabSingle5);
-
-		this.hideItemAll(LOTRMod.slabDouble);
-		this.hideItemAll(LOTRMod.slabDouble2);
-		this.hideItemAll(LOTRMod.slabDouble3);
-		this.hideItemAll(LOTRMod.slabDouble4);
-		this.hideItemAll(LOTRMod.slabDouble5);
-		this.hideItemAll(LOTRMod.slabDouble6);
-		this.hideItemAll(LOTRMod.slabDouble7);
-		this.hideItemAll(LOTRMod.slabDouble8);
-		this.hideItemAll(LOTRMod.slabDouble9);
-		this.hideItemAll(LOTRMod.slabDouble10);
-		this.hideItemAll(LOTRMod.slabDouble11);
-		this.hideItemAll(LOTRMod.slabDouble12);
-		this.hideItemAll(LOTRMod.slabDoubleV);
-		this.hideItemAll(LOTRMod.slabDoubleThatch);
-		this.hideItemAll(LOTRMod.slabDoubleDirt);
-		this.hideItemAll(LOTRMod.slabDoubleSand);
-		this.hideItemAll(LOTRMod.slabDoubleGravel);
-		this.hideItemAll(LOTRMod.rottenSlabDouble);
-		this.hideItemAll(LOTRMod.scorchedSlabDouble);
-		this.hideItemAll(LOTRMod.slabUtumnoDouble);
-		this.hideItemAll(LOTRMod.slabUtumnoDouble2);
-		this.hideItemAll(LOTRMod.slabClayTileDouble);
-		this.hideItemAll(LOTRMod.slabClayTileDyedDouble);
-		this.hideItemAll(LOTRMod.slabClayTileDyedDouble2);
-		this.hideItemAll(LOTRMod.woodSlabDouble);
-		this.hideItemAll(LOTRMod.woodSlabDouble2);
-		this.hideItemAll(LOTRMod.woodSlabDouble3);
-		this.hideItemAll(LOTRMod.woodSlabDouble4);
-		this.hideItemAll(LOTRMod.woodSlabDouble5);
+	private void registerHiddenItems() {
+		for (ItemHidingHandler handler : this.itemHidingHandlers) {
+			Collection<ItemStack> hiddenStacks = handler.getHiddenStacks();
+			if (hiddenStacks != null) {
+				hiddenStacks.forEach(API::hideItem);
+				NEIExtensions.mod.getLogger().debug(
+						"The item hiding handler \"" + handler.getClass() + "\" has hidden " + (hiddenStacks == null ? 0 : hiddenStacks.size()) + " items");
+			}
+		}
 	}
 
 	private void addOverrideNames() {
@@ -238,34 +150,6 @@ public class NEIIntegrationManager {
 		// Others
 		this.setOverrideName(LOTRMod.rhunFire, StatCollector.translateToLocal("neiLotr.lotr.block.rhunFire.name"));
 	}
-
-	private void hideItemMeta(ItemStack stack) {
-		for (int i = 8; i < 16; i++) {
-			ItemStack s = new ItemStack(stack.getItem(), 1, i);
-			API.hideItem(s);
-		}
-	}
-
-	private void hideItemMeta(Block block) {
-		this.hideItemMeta(new ItemStack(block));
-	}
-
-	// private void hideItemMeta(Item item) {
-	// this.hideItemMeta(new ItemStack(item));
-	// }
-
-	private void hideItemAll(ItemStack stack) {
-		ItemStack s = new ItemStack(stack.getItem(), 1, 32767);
-		API.hideItem(s);
-	}
-
-	private void hideItemAll(Block block) {
-		this.hideItemAll(new ItemStack(block));
-	}
-
-	// private void hideItemAll(Item item) {
-	// this.hideItemAll(new ItemStack(item));
-	// }
 
 	private void setOverrideName(ItemStack stack, String name) {
 		API.setOverrideName(stack, name);
